refactor: 특정 게시글 조회 응답 DTO 분리

+++ src/main/java/com/Project/BoardService/domain/entity/like/LikeRepository.java
@@ -14,5 +14,5 @@ public interface LikeRepository extends JpaRepository<PostLike, Long> {
 
     void deleteLikeByUserAndPost(User user, Post post);
 
-    Optional<Long> countLikeByPost(Post post);
+    Optional<Integer> countLikeByPost(Post post);
 }
